+++ compiler/rustc_middle/src/util/mod.rs
@@ -1,8 +1,5 @@
 pub mod bug;
 pub mod common;
-pub mod find_self_call;
-
-pub use find_self_call::find_self_call;
 
 #[derive(Default, Copy, Clone)]
 pub struct Providers {
